Move heroes app to test project
