+++ tests/test_admin.py
@@ -8,8 +8,8 @@ from django.test import RequestFactory
 from rest_framework_api_key.admin import APIKeyModelAdmin
 from rest_framework_api_key.models import APIKey
 
-from .heroes.admin import HeroAPIKeyModelAdmin
-from .heroes.models import Hero, HeroAPIKey
+from .project.heroes.admin import HeroAPIKeyModelAdmin
+from .project.heroes.models import Hero, HeroAPIKey
 
 
 @pytest.fixture(name="req")
